3_email_bot/requirements.txt
openai
chromadb
langchain-openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
beautifulsoup4
python-dotenv
